+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [3.13.0] - 2021-08-04
+### Added
+- Validation for criteria.
+### Changed
+- Disabled download button until the resource table data has been loaded.
+### Fixed
+- Slow evaluation of the contents of the resource table cells.
+
 ## [3.12.0] - 2021-07-29
 ### Changed
 - Observation search parameter of type "Quantity" now uses the composite